# onboard-service/pkg/onboard/types.gen.go
// Package onboard provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.1 DO NOT EDIT.
package onboard

import (
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// CardOnboardingStatusResponse defines model for CardOnboardingStatusResponse.
type CardOnboardingStatusResponse struct {
	AccountOnboardingStatus    string `json:"accountOnboardingStatus"`
	CustomerId                 string `json:"customerId"`
	CustomerRegistrationStatus string `json:"customerRegistrationStatus"`
	InterestDetailsStatus      string `json:"interestDetailsStatus"`
	OverallStatus              string `json:"overallStatus"`
}

// ErrorDetail defines model for ErrorDetail.
type ErrorDetail struct {
	Field string `json:"field"`
	Issue string `json:"issue"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code          string         `json:"code"`
	CorrelationId string         `json:"correlationId"`
	Details       *[]ErrorDetail `json:"details,omitempty"`
	Message       string         `json:"message"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// OnboardCardRequest defines model for OnboardCardRequest.
type OnboardCardRequest struct {
	CardNumber    string              `json:"cardNumber"`
	CardType      string              `json:"cardType"`
	CorrelationId string              `json:"correlationId"`
	CustomerId    string              `json:"customerId"`
	Email         openapi_types.Email `json:"email"`
	ExpiryDate    string              `json:"expiryDate"`
	HolderName    string              `json:"holderName"`
	JobId         string              `json:"jobId"`
	RecordId      string              `json:"recordId"`
	RowNumber     int32               `json:"rowNumber"`
	SourceFile    string              `json:"sourceFile"`
}

// OnboardCardResponse defines model for OnboardCardResponse.
type OnboardCardResponse struct {
	AccountId      string `json:"accountId"`
	CardId         string `json:"cardId"`
	CoreCustomerId string `json:"coreCustomerId"`
	CustomerId     string `json:"customerId"`
	Status         string `json:"status"`
}

// CorrelationIdHeader defines model for CorrelationIdHeader.
type CorrelationIdHeader = string

// OnboardCardParams defines parameters for OnboardCard.
type OnboardCardParams struct {
	XCorrelationId *CorrelationIdHeader `json:"X-Correlation-Id,omitempty"`
}

// GetCardOnboardingStatusParams defines parameters for GetCardOnboardingStatus.
type GetCardOnboardingStatusParams struct {
	XCorrelationId *CorrelationIdHeader `json:"X-Correlation-Id,omitempty"`
}

// OnboardCardJSONRequestBody defines body for OnboardCard for application/json ContentType.
type OnboardCardJSONRequestBody = OnboardCardRequest